feat(drtools): extract framework-agnostic dynamic-tools core into drtools.dynamic
global-mcp and other non-FastMCP/non-DrMCP hosts cannot import the dynamic
external-tool registrar today because it lives under drmcp, which transitively
imports `datarobot.context` (not present in the pinned `datarobot==3.14.0`
shipped with global-mcp). This change extracts the framework-agnostic core into
a new `datarobot_genai.drtools.dynamic` subpackage so any MCP host can reuse it.

What moved (no behavior change):
- `drmcp/core/dynamic_tools/schema.py` -> `drtools/dynamic/schema.py`
- `drmcp/core/dynamic_tools/deployment/adapters/{base,default,drum}.py` ->
  `drtools/dynamic/adapters/{base,default,drum}.py`
- `drmcp/core/dynamic_tools/deployment/schemas/*.json` ->
  `drtools/dynamic/schemas/*.json`
- `ExternalToolRegistrationConfig`, `_external_tool_callable_factory` (now public
  `build_external_tool_callable`), `get_outbound_headers`, the `REQUEST_*`
  constants and `REQUEST_FORWARDED_HEADERS` whitelist now live in
  `drtools/dynamic/external_tool.py`.

Host-specific bits (response formatting, ToolError class, FastMCP request
header extraction) are injected as parameters into
`build_external_tool_callable` so drtools.dynamic stays free of
fastmcp/drmcp imports (per `subpackages.import-restrictions`).

What stayed (server-bound shim):
- `drmcp/core/dynamic_tools/register.py` keeps `register_external_tool` as the
  drmcp public entry point. It now delegates to
  `build_external_tool_callable`, injecting drmcp's
  `format_response_as_tool_result`, `fastmcp.exceptions.ToolError`, and a
  drmcp-flavored `get_outbound_headers` (so existing tests can still
  monkeypatch `register.get_http_headers`).
- All deployment lifecycle modules (`deployment/{config,metadata,controllers,register}.py`)
  remain in drmcp — they depend on the drmcp `mcp` singleton, lineage manager,
  feature flags and SDK clients.

API guarantee:
All `from datarobot_genai.drmcp.core.dynamic_tools[.deployment.adapters].*`
imports continue to work — the old module paths are now thin re-export shims.

Tests:
- All 1310 drmcp unit tests pass; the 208 dynamic_tools tests are unchanged.
- New direct-import tests under `tests/drtools/unit/dynamic/` (45 tests) cover
  schema validation, adapters and the new `build_external_tool_callable`
  injection contract.
- Coverage on `datarobot_genai.drtools.dynamic` is 97.88% (well above 90%).

Pre-existing failures in `tests/drmcp/acceptance/` are out of scope (they
require live DataRobot API access and were red on `main` before this PR).

Follow-up PR 2 (in `datarobot/global-mcp`) will wire this into the global-mcp
FastMCP server behind a feature flag.
