+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "datarobot-genai"
-version = "0.15.41"
+version = "0.15.42"
 description = "Generic helpers for GenAI"
 readme = "README.md"
 requires-python = ">=3.11, <3.14"
@@ -39,7 +39,7 @@ where = ["src"]
 [tool.setuptools.package-data]
 datarobot_genai = [
     "py.typed",
-    "drmcp/core/dynamic_tools/deployment/schemas/*.json",
+    "drtools/dynamic/schemas/*.json",
 ]
 
 [tool.uv]
