+++ src/datarobot_genai/drtools/dynamic/schemas/drum_prediction_fallback_schema.json
